rename legendary bow to talent
